# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"authentication-controller","description":"Authentication Controller"},{"name":"baguette-item-controller","description":"Baguette Item Controller"},{"name":"baguette-order-controller","description":"Baguette Order Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"customer-controller","description":"Customer Controller"},{"name":"employee-controller","description":"Employee Controller"},{"name":"employee-role-controller","description":"Employee Role Controller"},{"name":"home-controller","description":"Home Controller"},{"name":"ingredient-controller","description":"Ingredient Controller"},{"name":"ingredient-type-controller","description":"Ingredient Type Controller"},{"name":"item-controller","description":"Item Controller"},{"name":"operation-handler","description":"Operation Handler"},{"name":"special-offer-controller","description":"Special Offer Controller"},{"name":"web-mvc-links-handler","description":"Web Mvc Links Handler"}],"paths":{"/":{"get":{"tags":["home-controller"],"summary":"getHello","operationId":"getHelloUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator":{"get":{"tags":["web-mvc-links-handler"],"summary":"links","operationId":"linksUsingGET","produces":["application/json","application/vnd.spring-boot.actuator.v2+json","application/vnd.spring-boot.actuator.v3+json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/definitions/Link"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/mappings":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET","produces":["application/json","application/vnd.spring-boot.actuator.v2+json","application/vnd.spring-boot.actuator.v3+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/all":{"get":{"tags":["baguette-item-controller"],"summary":"getBaguetteItems","operationId":"getBaguetteItemsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteItem"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/allByBaguetteOrder/{baguetteOrderId}":{"get":{"tags":["baguette-item-controller"],"summary":"getBaguetteItemsByBaguetteOrder","operationId":"getBaguetteItemsByBaguetteOrderUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteItem"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/detail/{baguetteItemId}":{"get":{"tags":["baguette-item-controller"],"summary":"getBaguetteItem","operationId":"getBaguetteItemUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteItemId","in":"path","description":"baguetteItemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteItem"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/all":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrders","operationId":"getBaguetteOrdersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteOrder"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/allByCustomer/{customerId}":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrdersByCustomer","operationId":"getBaguetteOrdersByCustomerUsingGET","produces":["*/*"],"parameters":[{"name":"customerId","in":"path","description":"customerId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteOrder"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/allByState/{state}":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrdersByState","operationId":"getBaguetteOrdersByStateUsingGET","produces":["*/*"],"parameters":[{"name":"state","in":"path","description":"state","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteOrder"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/detail/{baguetteOrderId}":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrder","operationId":"getBaguetteOrderUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteOrder"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/update/{baguetteOrderId}":{"put":{"tags":["baguette-order-controller"],"summary":"updateBaguetteOrder","operationId":"updateBaguetteOrderUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"},{"name":"state","in":"query","description":"state","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/basicauth":{"get":{"tags":["authentication-controller"],"summary":"helloWorldBean","operationId":"helloWorldBeanUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AuthToSend"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/customer/all":{"get":{"tags":["customer-controller"],"summary":"getUsers","operationId":"getUsersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Customer"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/customer/detail/{userId}":{"get":{"tags":["customer-controller"],"summary":"getUser","operationId":"getUserUsingGET","produces":["*/*"],"parameters":[{"name":"userId","in":"path","description":"userId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Customer"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/employee/admin/create":{"post":{"tags":["employee-controller"],"summary":"registerNewAdmin","operationId":"registerNewAdminUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"employee","description":"employee","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/employee/all":{"get":{"tags":["employee-controller"],"summary":"getEmployees","operationId":"getEmployeesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Employee"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/employee/create":{"post":{"tags":["employee-controller"],"summary":"registerNewEmployee","operationId":"registerNewEmployeeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"employee","description":"employee","required":true,"schema":{"$ref":"#/definitions/Employee"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/employee/detail/{employeeId}":{"get":{"tags":["employee-controller"],"summary":"getEmployee","operationId":"getEmployeeUsingGET","produces":["*/*"],"parameters":[{"name":"employeeId","in":"path","description":"employeeId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Employee"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/employee/remove/{employeeId}":{"delete":{"tags":["employee-controller"],"summary":"deleteEmployee","operationId":"deleteEmployeeUsingDELETE","produces":["*/*"],"parameters":[{"name":"employeeId","in":"path","description":"employeeId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/employee/update/{employeeId}":{"put":{"tags":["employee-controller"],"summary":"updateEmployee","operationId":"updateEmployeeUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"employeeId","in":"path","description":"employeeId","required":true,"type":"integer","format":"int64"},{"name":"firstname","in":"query","description":"firstname","required":false,"type":"string"},{"name":"lastname","in":"query","description":"lastname","required":false,"type":"string"},{"name":"login","in":"query","description":"login","required":false,"type":"string"},{"name":"password","in":"query","description":"password","required":false,"type":"string"},{"name":"roleId","in":"query","description":"roleId","required":false,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingGET","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingHEAD","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPOST","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPUT","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingDELETE","produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingOPTIONS","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"errorHtml","operationId":"errorHtmlUsingPATCH","consumes":["application/json"],"produces":["text/html"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ModelAndView"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/ingredient/all":{"get":{"tags":["ingredient-controller"],"summary":"getIngredients","operationId":"getIngredientsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Ingredient"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredient/create":{"post":{"tags":["ingredient-controller"],"summary":"createNewIngredient","operationId":"createNewIngredientUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"ingredient","description":"ingredient","required":true,"schema":{"$ref":"#/definitions/Ingredient"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredient/detail/{ingredientName}":{"get":{"tags":["ingredient-controller"],"summary":"getIngredientByName","operationId":"getIngredientByNameUsingGET","produces":["*/*"],"parameters":[{"name":"ingredientName","in":"path","description":"ingredientName","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Ingredient"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredient/remove/{ingredientId}":{"delete":{"tags":["ingredient-controller"],"summary":"deleteIngredient","operationId":"deleteIngredientUsingDELETE","produces":["*/*"],"parameters":[{"name":"ingredientId","in":"path","description":"ingredientId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/ingredient/update/{ingredientId}":{"put":{"tags":["ingredient-controller"],"summary":"updateIngredient","operationId":"updateIngredientUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"ingredientId","in":"path","description":"ingredientId","required":true,"type":"integer","format":"int64"},{"name":"ingredientTypeId","in":"query","description":"ingredientTypeId","required":false,"type":"integer","format":"int64"},{"name":"name","in":"query","description":"name","required":false,"type":"string"},{"name":"price","in":"query","description":"price","required":false,"type":"number","format":"double"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredientType/all":{"get":{"tags":["ingredient-type-controller"],"summary":"getIngredientTypes","operationId":"getIngredientTypesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/IngredientType"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredientType/create":{"post":{"tags":["ingredient-type-controller"],"summary":"createNewIngredientType","operationId":"createNewIngredientTypeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"ingredientType","description":"ingredientType","required":true,"schema":{"$ref":"#/definitions/IngredientType"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredientType/remove/{ingredientTypeId}":{"delete":{"tags":["ingredient-type-controller"],"summary":"deleteIngredientType","operationId":"deleteIngredientTypeUsingDELETE","produces":["*/*"],"parameters":[{"name":"ingredientTypeId","in":"path","description":"ingredientTypeId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/ingredientType/update/{ingredientTypeId}":{"put":{"tags":["ingredient-type-controller"],"summary":"updateIngredientType","operationId":"updateIngredientTypeUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"ingredientTypeId","in":"path","description":"ingredientTypeId","required":true,"type":"integer","format":"int64"},{"name":"name","in":"query","description":"name","required":false,"type":"string"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/all":{"get":{"tags":["item-controller"],"summary":"getItems","operationId":"getItemsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Item"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/create/{specialOfferId}":{"post":{"tags":["item-controller"],"summary":"addToSpecialOffer","operationId":"addToSpecialOfferUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"item","description":"item","required":true,"schema":{"$ref":"#/definitions/Item"}},{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Item"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/detail/{itemId}":{"get":{"tags":["item-controller"],"summary":"getItem","operationId":"getItemUsingGET","produces":["*/*"],"parameters":[{"name":"itemId","in":"path","description":"itemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Item"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/remove/{specialOfferId}/{itemId}":{"delete":{"tags":["item-controller"],"summary":"removeItem","operationId":"removeItemUsingDELETE","produces":["*/*"],"parameters":[{"name":"itemId","in":"path","description":"itemId","required":true,"type":"integer","format":"int64"},{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/item/update/{specialOfferId}/{itemId}":{"put":{"tags":["item-controller"],"summary":"updateItem","operationId":"updateItemUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"amount","in":"query","description":"amount","required":false,"type":"integer","format":"int32"},{"name":"itemId","in":"path","description":"itemId","required":true,"type":"integer","format":"int64"},{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/{baguetteItemId}/all":{"get":{"tags":["item-controller"],"summary":"getItemsByBaguetteItem","operationId":"getItemsByBaguetteItemUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteItemId","in":"path","description":"baguetteItemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Item"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/{specialOfferId}/all":{"get":{"tags":["item-controller"],"summary":"getItemsBySpecialOffer","operationId":"getItemsBySpecialOfferUsingGET","produces":["*/*"],"parameters":[{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Item"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/all":{"get":{"tags":["employee-role-controller"],"summary":"getEmployeeRoles","operationId":"getEmployeeRolesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/EmployeeRole"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/create":{"post":{"tags":["employee-role-controller"],"summary":"registerNewEmployeeRole","operationId":"registerNewEmployeeRoleUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"employeeRole","description":"employeeRole","required":true,"schema":{"$ref":"#/definitions/EmployeeRole"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/role/detail/{employeeRoleId}":{"get":{"tags":["employee-role-controller"],"summary":"getEmployeeRole","operationId":"getEmployeeRoleUsingGET","produces":["*/*"],"parameters":[{"name":"employeeRoleId","in":"path","description":"employeeRoleId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/EmployeeRole"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/all":{"get":{"tags":["special-offer-controller"],"summary":"getSpecialOffers","operationId":"getSpecialOffersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SpecialOffer"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/all/{active}":{"get":{"tags":["special-offer-controller"],"summary":"getSpecialOffersByActive","operationId":"getSpecialOffersByActiveUsingGET","produces":["*/*"],"parameters":[{"name":"active","in":"path","description":"active","required":true,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SpecialOffer"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/create":{"post":{"tags":["special-offer-controller"],"summary":"createNewSpecialOffer","operationId":"createNewSpecialOfferUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"specialOffer","description":"specialOffer","required":true,"schema":{"$ref":"#/definitions/SpecialOffer"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SpecialOffer"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/detail/{specialOfferId}":{"get":{"tags":["special-offer-controller"],"summary":"getSpecialOffer","operationId":"getSpecialOfferUsingGET","produces":["*/*"],"parameters":[{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SpecialOffer"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/update/{specialOfferId}":{"put":{"tags":["special-offer-controller"],"summary":"updateSpecialOffer","operationId":"updateSpecialOfferUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"active","in":"query","description":"active","required":false,"type":"boolean"},{"name":"items","in":"query","description":"items","required":false,"type":"string"},{"name":"name","in":"query","description":"name","required":false,"type":"string"},{"name":"price","in":"query","description":"price","required":false,"type":"number","format":"double"},{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/updateActivity/{specialOfferId}":{"put":{"tags":["special-offer-controller"],"summary":"updateActivitySpecialOffer","operationId":"updateActivitySpecialOfferUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"active","in":"query","description":"active","required":false,"type":"boolean"},{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"AuthToSend":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roleName":{"type":"string"}},"title":"AuthToSend"},"BaguetteItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/definitions/Item"}},"offer":{"type":"boolean"},"price":{"type":"number","format":"double"}},"title":"BaguetteItem"},"BaguetteOrder":{"type":"object","properties":{"customer":{"$ref":"#/definitions/Customer"},"date":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"note":{"type":"string"},"price":{"type":"number","format":"double"},"state":{"type":"integer","format":"int32"}},"title":"BaguetteOrder"},"Customer":{"type":"object","properties":{"email":{"type":"string"},"emailShort":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"Customer"},"Employee":{"type":"object","properties":{"employeeRole":{"$ref":"#/definitions/EmployeeRole"},"firstname":{"type":"string"},"id":{"type":"integer","format":"int64"},"lastname":{"type":"string"},"login":{"type":"string"},"password":{"type":"string"}},"title":"Employee"},"EmployeeRole":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"EmployeeRole"},"Ingredient":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ingredientType":{"$ref":"#/definitions/IngredientType"},"name":{"type":"string"},"price":{"type":"number","format":"double"}},"title":"Ingredient"},"IngredientType":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"IngredientType"},"Item":{"type":"object","properties":{"amount":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"},"ingredient":{"$ref":"#/definitions/Ingredient"},"name":{"type":"string"},"price":{"type":"number","format":"double"}},"title":"Item"},"Link":{"type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}},"title":"Link"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"SpecialOffer":{"type":"object","properties":{"active":{"type":"boolean"},"id":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/definitions/Item"}},"name":{"type":"string"},"price":{"type":"number","format":"double"}},"title":"SpecialOffer"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}